+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0]
+
+### Added
+
+* Link#of now accepts URIs with templated query parameters
+
+### Removed
+
+* Removed Spring Boot org.springframework.boot:spring-boot-starter-webflux. Only actually used libraries are
+  incorporated:
+    * org.springframework:spring-webflux
+    * org.springframework.boot:spring-boot-starter-json
+
 ## [0.3.0]
 
 ### Added